move mapbox to own component
